Introduce guys to S², delete Solidity compiler, &c
I also deleted the S³ prototype, because it will be different.

This new S² is not correctly compiled yet, because it doesn't
actually do any of the authority checks.

I've also introduced a bunch of Haskell FFI stuff in order to
do Keccak-256 hashing in Agda, and (as a proof-of-concept) also
read environment variables.

It seems reasonable to actually make the deploy scripts themselves
in Agda now, because we can easily imagine a blocking I/O postulate
that does, e.g., `seth create` via the Haskell system interface...